Firebase deploy related problem solved

+++ src/Components/CartProductCard.jsx
@@ -1,28 +1,47 @@
 import { useState } from "react";
+import Swal from 'sweetalert2'
 
 const CartProductCard = ({ product }) => {
     const { _id, name, image, brand, price, rating } = product;
     // console.log(product);
-    const [ state, setState ] = useState(true);
+    const [state, setState] = useState(true);
     const handleDeleteProduct = () => {
-        fetch(`https://brand-shop-server-rho-seven.vercel.app/myCart/${_id}`, {
-            method: "DELETE",
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonText: 'Yes, delete it!',
+            cancelButtonText: 'No, cancel!',
+            reverseButtons: true
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`https://brand-shop-server-rho-seven.vercel.app/myCart/${_id}`, {
+                    method: "DELETE",
+                })
+                    .then(res => res.json())
+                    .then(() => {
+                        // if (data.deletedCount > 0) {
+                        //     Swal.fire({
+                        //         title: 'Product Deleted!',
+                        //         text: 'Enjoy Exploring!',
+                        //         icon: 'success',
+                        //         confirmButtonText: 'Continue'
+                        //     });
+                        // }
+                        setState(false);
+                    })
+                    .catch(error => console.error(error))
+            }
         })
-            .then(res => res.json())
-            .then(data => {
-                if (data.deletedCount > 0) {
-                    alert("Deleted successfully")
-                }
-                setState(false);
-            })
-            .catch(error => console.error(error))
+
 
     }
     console.log(state);
 
     return (
         <div>
-            { state&&
+            {state &&
                 <div className="w-full h-36 bg-slate-100 rounded-lg flex justify-center items-center">
                     <div className="flex justify-around items-center w-full">
                         <img className="w-[80px]" src={image} alt="hero" />
